requirements.txt
# orm
sqlmodel
# driver for postgres
psycopg2-binary
# data security
python-decouple
# API Framework
fastapi[all]